move modal route registration to the route module
