+++ src/packages/core/modal/context/index.ts
@@ -1,3 +1,3 @@
-export type { UmbModalContext, UmbModalRejectReason } from './modal.context.js';
+export type { UmbModalContext, UmbModalRejectReason, UmbModalContextClassArgs } from './modal.context.js';
 export { UMB_MODAL_CONTEXT } from './modal.context-token.js';
 export * from './modal-manager.context.js';
