+++ src/packages/core/modal/index.ts
@@ -1,8 +1,6 @@
 import './component/modal.element.js';
 
 export * from './context/index.js';
-export * from './route-registration/modal-route-registration.interface.js';
-export * from './route-registration/modal-route-registration.controller.js';
 export * from './token/index.js';
 export * from './types.js';
 export * from './component/modal-base.element.js';
